+++ EntityFrameworkCore.Data/FootballLeagueDbContext.cs
@@ -3,6 +3,7 @@ using EntityFrameworkCore.Domain;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.Logging;
 using System.Reflection;
+using EntityFrameworkCore.Data.Utility;
 // using EntityFrameworkCore.Data.Configurations;
 
 namespace EntityFrameworkCore.Data;
@@ -17,6 +18,7 @@ public class FootballLeagueDbContext : DbContext
 
     protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
     {
+        EnvironmentVariableUtility.LoadEnv();
         var password = Environment.GetEnvironmentVariable("DB_PASSWORD")
             ?? throw new InvalidOperationException("Environment variable DB_PASSWORD is not set.");
